fix: address 6 critical CodeRabbit issues
- AuctionLifecycleService: withdraw from winner before state change; persist before seller payout
- RentScheduler: use config collectionInterval instead of hard-coded 60s
- ShopCreateListener: default-deny GUILD owner access; remove premature ShopCreatedEvent
- ContainerTradeService: check removeItem leftovers; validate rollback compensation
- V003→V006: migrate timestamp columns from INTEGER to BIGINT (epoch millis overflow)
- ShopTradeService: add compensation failure logging; validate positive price

+++ src/main/kotlin/net/badgersmc/em/infrastructure/persistence/AuctionRepositorySql.kt
@@ -67,7 +67,7 @@ class AuctionRepositorySql(private val ds: DataSource) : AuctionRepository {
                 } else {
                     ps.setNull(5, java.sql.Types.INTEGER)
                     ps.setNull(6, java.sql.Types.VARCHAR)
-                    ps.setNull(7, java.sql.Types.INTEGER)
+                    ps.setNull(7, java.sql.Types.BIGINT)
                 }
                 ps.setLong(8, auction.antiSnipeWindow.toSeconds())
                 ps.setString(9, auction.id.value)
@@ -106,7 +106,7 @@ class AuctionRepositorySql(private val ds: DataSource) : AuctionRepository {
         } else {
             ps.setNull(7, java.sql.Types.INTEGER)
             ps.setNull(8, java.sql.Types.VARCHAR)
-            ps.setNull(9, java.sql.Types.INTEGER)
+            ps.setNull(9, java.sql.Types.BIGINT)
         }
         ps.setLong(10, auction.antiSnipeWindow.toSeconds())
     }
